% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_suffix.R
\name{as_suffix}
\alias{as_suffix}
\title{Convert a string into a suffix}
\usage{
as_suffix(x, default = "")
}
\arguments{
\item{x}{string to format as suffix}

\item{default}{what to return when x is null. Defaults to ""}
}
\value{
A string
}
\description{
Generally, this simply means add a "_" to the beginning of your string.
But it's main purpose is to use inside other functions,
specifically taking function args as input, which might be NULL (default)
and as such we want to return a dummy suffix of "".
}
\details{
It means that instead of calling a function with args like this:
\itemize{
\item pull_data(version = "_22")
}

I can call a function with args like this:
\itemize{
\item pull_data(version = "22")
}

Because inside pull_data I'm doing:

version <- version |> as_suffix()

Which is appending "_" at the beginning, or, if NULL, not doing anything.
}
\examples{
\dontrun{
as_suffix(NULL)
as_suffix("22")
as_suffix(22)
as_suffix(NULL, "_Master")
}
}
