% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_yyyymmdd_hhmmss.R
\name{as_yyyymmdd_hhmmss}
\alias{as_yyyymmdd_hhmmss}
\title{System Date and Time in YYYYMMDD_HHMMSS format}
\usage{
as_yyyymmdd_hhmmss(x = Sys.time())
}
\arguments{
\item{x}{the date-time to be formatted as YYYYMMDD_HHMMSS, and defaulting to Sys.time() the function is called}
}
\value{
a \if{html}{\out{<chr>}} value of the current date-time in YYYYMMDD_HHMMSS format
}
\description{
System Date and Time in YYYYMMDD_HHMMSS format
}
\examples{
as_yyyymmdd_hhmmss()
t0 <- Sys.time()
as_yyyymmdd_hhmmss(t0)
}
