% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_nfu.R
\name{connect_nfu}
\alias{connect_nfu}
\title{Connect to database via ODBC}
\usage{
connect_nfu(dsn = "UWPSQL06", server = dsn, database = "Sandbox", ...)
}
\arguments{
\item{dsn}{The Data Source Name. A character scalar.}

\item{server}{The server name. A character scalar.}

\item{database}{The database name. A character scalar.}

\item{...}{}
}
\value{
The output of dbConnect: I expect the class "Microsoft SQL Server"
}
\description{
A convenience wrapper around DBI::dbConnect().
Usefully for Pricing, it defaults to the UWPSQL06 server
}
\details{
NB ODBC Data Sources must be set up correctly.
Basically, add a User DSN for each server.
I used the driver SQL Server Native Client 11.0.
Name them after the server for the defaults to work nicely.
You can use the same name for the Description.
For example, add a DSN called "UWPSQL06" for the server "UWPSQL06".
}
\examples{
conn <- connect_nfu()
conn_6 <- connect_nfu("UWPSQL06")
conn_4 <- connect_nfu("UWPSQL04")
conn_2 <- connect_nfu("UWPSQL02")
}
