% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw_crime_data.R
\name{process_raw_crime_data}
\alias{process_raw_crime_data}
\title{Process raw crime data and left join onto postcode data (OSPCode)}
\usage{
process_raw_crime_data(version_crime = NULL, conn = connect_nfu())
}
\arguments{
\item{version_crime}{\emph{eg} "20220923".
Default (NULL) means we pull from the table with no suffix}

\item{conn}{Defaults to connect_nfu(), \emph{ie} the default server.}
}
\value{
lazy tibble
}
\description{
Process raw crime data and left join onto postcode data (OSPCode)
}
\examples{
\doontrun{
process_raw_crime_data()
process_raw_crime_data("20220923")
}
}
