% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ospcode.R
\name{read_ospcode}
\alias{read_ospcode}
\title{Read OSPCode}
\usage{
read_ospcode(version_postcodes = NULL, conn = connect_nfu())
}
\arguments{
\item{version_postcodes}{\emph{eg} "2206".
Default (NULL) creates a suffix of "_Master"}

\item{conn}{Defaults to connect_nfu(), \emph{ie} the default server.}
}
\value{
lazy tibble
}
\description{
Specify the table version you want. By default, reads OSPCode_Master.
}
\examples{
read_ospcode() # reads OSPCode_Master
read_ospcode("2206") # reads OSPCode_2206
}
