% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_aadf.R
\name{scrape_aadf}
\alias{scrape_aadf}
\title{Scrape the latest AADF data from the government website}
\usage{
scrape_aadf(raw_counts = FALSE)
}
\arguments{
\item{raw_counts}{By default, FALSE, which means it scrapes "dft_traffic_counts_aadf".
If set to TRUE, then it scrapes "dft_traffic_counts_raw_counts" instead.}
}
\value{
A tibble
}
\description{
Scrape the latest AADF data from the government website
}
\examples{
\dontrun{
aadf <- scrape_aadf()
aadf <- scrape_aadf(raw_counts = TRUE)
}
}
