% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_road_safety_data.R
\name{scrape_road_safety_data}
\alias{scrape_road_safety_data}
\title{Scrape the latest Road Safety Data from the government website}
\usage{
scrape_road_safety_data(year = NULL, vehicle = FALSE)
}
\arguments{
\item{year}{The year of data to return. By default, NULL, the most recent year available.}

\item{vehicle}{If TRUE, then this scrapes "dft-road-casualty-statistics-vehicle" data
instead of "dft-road-casualty-statistics-collision" data. By default it is FALSE.}
}
\value{
A tibble containing the collisions (or vehicle) data
}
\description{
Also associated with (not sure exactly what the relationship is) the term "Stats19".
This is data on traffic collisions.
}
\examples{
\dontrun{
accidents <- scrape_road_safety_data()
vehicles <- scrape_road_safety_data(vehicle = TRUE)
accidents_2020 <- scrape_road_safety_data(2020)
vehicles_2020 <- scrape_road_safety_data(2020)
}
}
