% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_as_sf_27700_safe.R
\name{st_as_sf_27700_safe}
\alias{st_as_sf_27700_safe}
\title{Convert foreign object to an sf object with British National Grid, safely}
\usage{
st_as_sf_27700_safe(df, x, y)
}
\arguments{
\item{df}{data.frame}

\item{x}{variable name to use as x coord}

\item{y}{variable name to use as y coord}
}
\value{
An sf object
}
\description{
Convert foreign object to an sf object with British National Grid, safely
}
\examples{
\dontrun{
df |> st_as_sf("Eastings", "Northings")
}
}
