% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_name.R
\name{table_name}
\alias{table_name}
\title{Get prettified table name from conn (DBI::dbConnect) and Id (DBI::Id)}
\usage{
table_name(Id, conn = connect_nfu(), full_name = TRUE)
}
\arguments{
\item{Id}{the output of DBI::Id containing the schema and table name}

\item{conn}{the output of DBI::dbConnect giving the server and db}

\item{full_name}{Defaults to TRUE meaning server and db is also given}
}
\value{
a \if{html}{\out{<chr>}} name of the table of the format server.db.schema.table
}
\description{
By default, gives the fully qualified name with server and database specified
Optionally, if full_name is FALSE, it will just give the schema and table
}
\examples{
\dontrun{table_name(conn, Id)}
}
