% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThePricingEngineer-package.R
\docType{package}
\name{ThePricingEngineer-package}
\alias{ThePricingEngineer}
\alias{ThePricingEngineer-package}
\title{ThePricingEngineer: Preparing Data for Analysis and Modelling}
\description{
Data Engineering, Feature Engineering, Preparing Data for Analysis and Modelling by the teams in the Pricing and Underwriting Department. Initially designed for the Loss Cost Modelling (LCM) team
}
\author{
\strong{Maintainer}: Daniel Braddock \email{daniel_braddock@nfumutual.co.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
