% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_aadf.R
\name{scrape_aadf}
\alias{scrape_aadf}
\title{Scrape the latest AADF data from the government website}
\usage{
scrape_aadf(raw_counts = FALSE)
}
\arguments{
\item{raw_counts}{By default, FALSE, which means it scrapes "dft_traffic_counts_aadf".
If set to TRUE, then it scrapes "dft_traffic_counts_raw_counts" instead.}
}
\value{
A tibble containing the aadf (or raw_counts) data
}
\description{
Uses rvest to scrape the data location and httr2 to perform a request.
This saves a .zip file containing the dataset in a .csv file to a temp dir.
This data is then read in and returned.
}
\examples{
\dontrun{
aadf <- scrape_aadf()
aadf <- scrape_aadf(raw_counts = TRUE)
}
}
