% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThePricingEngineer-package.R
\docType{package}
\name{ThePricingEngineer-package}
\alias{ThePricingEngineer}
\alias{ThePricingEngineer-package}
\title{ThePricingEngineer: Preparing Data for Analysis and Modelling}
\description{
For interacting with NFU SQL servers, Data Engineering, Feature Engineering, and preparing data for analysis and modelling. For all NFU Pricing and Underwriting teams but initially designed for the Loss Cost Modelling (LCM) team.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DanielBraddock/ThePricingEngineer}
  \item Report bugs at \url{https://github.com/DanielBraddock/ThePricingEngineer/issues}
}

}
\author{
\strong{Maintainer}: Daniel Braddock \email{daniel_braddock@nfumutual.co.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
