% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sql.R
\name{write_sql}
\alias{write_sql}
\title{write_sql}
\usage{
write_sql(
  df,
  to_conn,
  to_schema,
  to_table,
  time_stamp = TRUE,
  overwrite = TRUE,
  temporary = FALSE
)
}
\arguments{
\item{df}{the local data.frame to write to SQL}

\item{to_conn}{a remote data source, eg SQL server + database created from DBI::dbConnect}

\item{to_schema}{the name of the schema to write to eg "this_schema"}

\item{to_table}{the name of the table to write to (needs no datetime stamp) eg "this_table"}

\item{time_stamp}{defaults to TRUE, creating a datetime stamp suffix for you table}

\item{overwrite}{defaults to TRUE, overwriting any existing table}

\item{temporary}{defaults to FALSE, ensuring the table persists after the sessions ends}
}
\description{
This function is called for it's side-effects of writing a local data.frame to SQL
}
\examples{
\dontrun{write_sql(df, conn, "this_schema", "this_table")}
}
